requirements.txt
asgiref==3.4.1
autopep8==1.5.7
black==21.8b0
click==8.0.1
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pycodestyle==2.7.0
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
